dwadaw
awdawdawd

+++ app/Models/VendorEarning.php
@@ -155,7 +155,7 @@ class VendorEarning extends Model
             'withholding_tax_amount' => round($withholdingTaxAmount, 2),
             'net_earning' => round($netEarning, 2),
             'earning_status' => 'pending',
-            'available_at' => now()->addDays(config('finance.settlement.auto_available_after_days', 7)),
+            'available_at' => now()->addDays((int) config('finance.settlement.auto_available_after_days', 7)),
         ]);
     }
